web-app/app.py
from flask import Flask, jsonify, send_file, request
from flask_cors import CORS
import gspread
from google.oauth2.service_account import Credentials
from urllib.parse import unquote
from datetime import datetime, date
import os
import json
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)
CORS(app)

SCOPES = ["https://www.googleapis.com/auth/spreadsheets"]
SHEET_ID = os.environ.get("SHEET_KEY", "")

CUSTOMER_COLUMNS = ["customer", "region", "sales_person", "customer_segment",
                    "customer_reference", "customer_number", "adress_number",
                    "postal_code", "phone", "email"]

ORDER_COLUMNS = ["Reference", "Order date", "Delivery date", "Customer", "Customer Reference",
                 "Buyer number", "Customer number", "Logistics number", "Address", "Number",
                 "Postal code", "City", "Country", "Phone number", "SKU", "Product", "Weight",
                 "Quantity", "Total weight", "Unit", "Total (Pre-discount)", "Product Discount",
                 "Total", "Currency", "Order Discount (Amount)", "Order Discount (%)", "Batch"]

CONTACT_COLUMNS = ["date_time", "sales_person", "customer", "contact_channel", "result",
                   "comment", "customer_contact_person", "follow_up_date"]


_spreadsheet_cache = None

def get_spreadsheet():
    global _spreadsheet_cache
    if _spreadsheet_cache is None:
        creds = Credentials.from_service_account_info(json.loads(os.environ["GOOGLE_CREDENTIALS"]), scopes=SCOPES)
        _spreadsheet_cache = gspread.authorize(creds).open_by_key(SHEET_ID)
    return _spreadsheet_cache


def rows_to_dicts(rows, columns):
    result = []
    for row in rows:
        padded = row + [""] * (len(columns) - len(row))
        result.append(dict(zip(columns, padded[:len(columns)])))
    return result


@app.route("/")
def index():
    return send_file("index.html")


@app.route("/customers", methods=["GET"])
def get_customers():
    spreadsheet = get_spreadsheet()
    sheet = spreadsheet.worksheet("customers_enriched")
    all_rows = sheet.get_all_values()
    headers = all_rows[0]

    # Build latest follow_up_date per customer from sales_activities
    contact_rows = rows_to_dicts(spreadsheet.worksheet("sales_activities").get_all_values()[1:], CONTACT_COLUMNS)
    latest_followup = {}
    for c in contact_rows:
        name = c["customer"].strip().lower()
        nf = c["follow_up_date"].strip()
        if nf and (name not in latest_followup or nf > latest_followup[name]):
            latest_followup[name] = nf

    def parse_coord(val):
        try:
            return float(val.replace(",", ".")) if val else None
        except ValueError:
            return None

    customers = []
    for i, row in enumerate(all_rows[1:], start=2):
        padded = row + [""] * (len(headers) - len(row))
        d = dict(zip(headers, padded))
        customer = {col: d.get(col, "") for col in CUSTOMER_COLUMNS}
        customer["latitude"]  = parse_coord(d.get("latitude_google") or d.get("latitude",  ""))
        customer["longitude"] = parse_coord(d.get("longitude_google") or d.get("longitude", ""))
        addr = d.get("address_google", "").strip()
        num  = d.get("address_number_google", "").strip()
        customer["address"] = f"{addr} {num}".strip()
        customer["city"] = d.get("city_google", "").strip() or d.get("city", "")
        customer["follow_up_date"] = latest_followup.get(customer["customer"].strip().lower(), "")
        customers.append({"row": i, **customer})
    return jsonify(customers)


@app.route("/customers/<customer_name>/stats", methods=["GET"])
def get_customer_stats(customer_name):
    customer_name = unquote(customer_name).strip().lower()
    spreadsheet = get_spreadsheet()

    # Orders
    order_rows = rows_to_dicts(spreadsheet.worksheet("order_rows").get_all_values()[1:], ORDER_COLUMNS)
    total_sales = 0.0
    latest_order_date = None
    currency = ""

    unique_references = set()
    for o in order_rows:
        if o["Customer"].strip().lower() != customer_name:
            continue
        try:
            cleaned = "".join(c for c in o["Total"] if c.isdigit() or c in ".,").replace(",", ".")
            if cleaned:
                total_sales += float(cleaned)
        except ValueError:
            pass
        if not currency and o["Currency"].strip():
            currency = o["Currency"].strip()
        d = o["Order date"].strip()
        if d and (latest_order_date is None or d > latest_order_date):
            latest_order_date = d
        if o["Reference"].strip():
            unique_references.add(o["Reference"].strip())

    # Contacts
    contact_rows = rows_to_dicts(spreadsheet.worksheet("sales_activities").get_all_values()[1:], CONTACT_COLUMNS)
    contacts = [
        {k: c[k] for k in ("customer", "date_time", "sales_person", "contact_channel", "result", "comment", "customer_contact_person", "follow_up_date")}
        for c in contact_rows
        if c["customer"].strip().lower() == customer_name
    ]
    contacts.sort(key=lambda x: x["date_time"], reverse=True)

    return jsonify({
        "total_sales": round(total_sales, 2),
        "latest_order_date": latest_order_date or "—",
        "currency": currency,
        "order_count": len(unique_references),
        "contacts": contacts,
    })


@app.route("/customer-insights", methods=["GET"])
def get_customer_insights():
    spreadsheet = get_spreadsheet()
    today = date.today()

    # Latest follow_up_date per customer
    contact_rows = rows_to_dicts(spreadsheet.worksheet("sales_activities").get_all_values()[1:], CONTACT_COLUMNS)
    latest_followup = {}
    for c in contact_rows:
        name = c["customer"].strip().lower()
        nf = c["follow_up_date"].strip()
        if nf and (name not in latest_followup or nf > latest_followup[name]):
            latest_followup[name] = nf

    # Latest order date and order count per customer
    order_rows = rows_to_dicts(spreadsheet.worksheet("order_rows").get_all_values()[1:], ORDER_COLUMNS)
    latest_order = {}
    latest_delivery = {}
    order_count = {}
    for o in order_rows:
        name = o["Customer"].strip().lower()
        d = o["Order date"].strip()
        dd = o["Delivery date"].strip()
        ref = o["Reference"].strip()
        if d and (name not in latest_order or d > latest_order[name]):
            latest_order[name] = d
        if dd and (name not in latest_delivery or dd > latest_delivery[name]):
            latest_delivery[name] = dd
        if ref:
            order_count[name] = order_count.get(name, 0) + 1

    # Compute insights for all customers
    all_names = set(latest_followup.keys()) | set(latest_order.keys()) | set(order_count.keys()) | set(latest_delivery.keys())
    insights = {}
    for name in all_names:
        # missad_uppfoljning
        nf = latest_followup.get(name, "")
        missad = bool(nf and nf < today.isoformat())

        # customer_risk
        lo = latest_order.get(name, "")
        count = order_count.get(name, 0)
        if count == 0 or not lo:
            risk = ""
        else:
            try:
                lo_date = date.fromisoformat(lo[:10])
                days = (today - lo_date).days
                if days > 60:
                    risk = "FÖRLORAD?"
                elif days > 40:
                    risk = "HÖG RISK"
                elif days > 20:
                    risk = "RISK"
                else:
                    risk = "OK"
            except ValueError:
                risk = ""

        ld = latest_delivery.get(name, "")
        insights[name] = {
            "missad_uppfoljning": missad,
            "customer_risk": risk,
            "latest_delivery_date": ld,
            "latest_delivery_month": ld[:7] if ld else "",  # "YYYY-MM"
        }

    return jsonify(insights)


@app.route("/customers/<customer_name>/contacts", methods=["POST"])
def add_contact(customer_name):
    customer_name = unquote(customer_name)
    data = request.get_json()
    sheet = get_spreadsheet().worksheet("sales_activities")
    row = [
        data.get("date_time", datetime.now().strftime("%Y-%m-%d %H:%M")),
        data.get("sales_person", ""),
        customer_name,
        data.get("contact_channel", ""),
        data.get("result", ""),
        data.get("comment", ""),
        data.get("customer_contact_person", ""),
        data.get("follow_up_date", ""),
    ]
    sheet.append_row(row)
    return jsonify({"ok": True})



if __name__ == "__main__":
    app.run(host="0.0.0.0", port=5000, debug=True)
